cast pointer to struct to long unsigned int to get rid of warning
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@13440 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/bbogart/popup/popup.c
@@ -639,7 +639,7 @@ static void *popup_new(t_symbol *s, int argc, t_atom *argv)
     }
 
       /* Bind the recieve "popup%lx" to the widget outlet*/
-      sprintf(buf,"popup%lx",x);
+      sprintf(buf,"popup%lx", (long unsigned int)x);
       x->x_sym = gensym(buf);
       pd_bind(&x->x_obj.ob_pd, x->x_sym);
